requirements_cleaned.txt
asgiref==3.9.1
cffi==2.0.0
cryptography==46.0.3
distlib==0.4.0
Django==4.2.25
django-extensions==4.1
djangorestframework==3.15.2
filelock==3.19.1
MarkupSafe==3.0.3
pillow==11.3.0
platformdirs==4.4.0
psycopg2-binary==2.9.10
pycparser==2.23
pyOpenSSL==25.3.0
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.34.0
Werkzeug==3.1.3
